pyproject.toml
[project]
name = "triton-unfold"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cupy-cuda12x>=13.3.0",
    "easydict>=1.13",
    "jaxtyping>=0.2.36",
    "numpy==1.22.4",
    "setuptools>=75.6.0",
    "sigpy>=0.1.26",
    "torch>=2.5.1",
    "triton==3.1",
]
